重构渲染系统以引入 RenderPipeline 和 RenderSystem
重构渲染逻辑，新增 RenderPipeline 和 RenderSystem 模块：
- RenderPipeline：负责渲染管线的初始化、执行和资源管理，支持阴影渲染、几何渲染和后期合成。
- RenderSystem：作为渲染系统管理类，集成 RenderPipeline，支持编辑器相机设置和视口调整。
- 新增 RenderConfig，统一管理光照、环境贴图和渲染配置。

其他改动：
- 移除 Scene 中的 Environment 和 Light 定义，迁移至 RenderConfig。
- 修改 OpenGLPipeline 和 Pipeline 的 Bind 方法为 const。
- 修复 PhysicsLayer 和 PXPhysicsWrappers 中的类型转换问题。
- Mesh.h 添加 RenderPipeline 友元声明。

+++ Prism/src/Prism/Physics/PXPhysicsWrappers.cpp
@@ -151,7 +151,7 @@ namespace Prism {
         std::vector<Vertex> vertices = collider.CollisionMesh->GetStaticVertices();
 
         physx::PxConvexMeshDesc convexDesc;
-        convexDesc.points.count = vertices.size();
+        convexDesc.points.count = (uint32_t)vertices.size();
         convexDesc.points.stride = sizeof(Vertex);
         convexDesc.points.data = vertices.data();
         convexDesc.flags = physx::PxConvexFlag::eCOMPUTE_CONVEX;
